+++ Nos/Views/Components/NoteTextEditor.swift
@@ -41,7 +41,7 @@ struct NoteTextEditor: View {
                     relatedAuthors: relatedAuthors,
                     emptyPlaceholder: { EmptyView() },
                     didSelectGesture: { [weak controller] author in
-                        /// Guard against double presses
+                        // Guard against double presses
                         guard let controller, controller.showMentionsAutocomplete else { return }
                         
                         controller.insertMention(of: author)